Rename componentCollector.ts -> componentCollectorSpaces.ts
