+++ components/experimental/spaces/src/interfaces/componentCollectorSpaces.ts
