+++ components/experimental/spaces/src/interfaces/index.ts
@@ -4,7 +4,7 @@
  */
 
 export * from "./componentCallable";
-export * from "./componentCollector";
+export * from "./componentCollectorSpaces";
 export * from "./componentToolbar";
 export * from "./componentToolbarConsumer";
 export * from "./componentRegistryDetails";
